+++ tests/test_pfctl_exporter.py
@@ -29,7 +29,7 @@ def test_parse_interfaces_output(caplog) -> None:
     assert "got 67 lines of output from 'pfctl -vvs Interfaces'" in caplog.text
     assert "found new interface: 'em0.3' - getting metrics..." in caplog.text
     assert (
-        'Adding new Gauge metric pfctl_interfaces_cleared_timestamp_seconds{interface="em0.7"} 1700416241'
+        'Adding new Gauge metric pfctl_interface_cleared_timestamp_seconds{interface="em0.7"} 1700416241'
         in caplog.text
     )
 
@@ -40,6 +40,6 @@ def test_parse_rules_output(caplog) -> None:
     for _ in c.collect_rules(mock_output=sample_rules_output.split("\n")):
         pass
     assert (
-        'Adding new Counter metric: pfctl_rules_bytes_total{rule="pass in quick on em0.7 proto tcp from <prometheus6> to <nuc2> port = 9999 flags S/SA keep state"} 33758359'
+        'Adding new Counter metric: pfctl_rule_bytes_total{rule="pass in quick on em0.7 proto tcp from <prometheus6> to <nuc2> port = 9999 flags S/SA keep state"} 33758359'
         in caplog.text
     )